@@ -1,14 +1,17 @@
 {
   "name": "kidsloop-live-server",
-  "version": "1.8.0",
+  "version": "1.10.4",
   "description": "",
   "main": "main.js",
   "scripts": {
-    "start": "ts-node dist/index.js",
-    "dev": "ts-node-dev src/main.ts",
-    "build": "webpack --stats-modules-space 999",
-    "lint": "eslint --fix .",
-    "test": "",
+    "start": "node dist/main.js",
+    "dev": "ts-node-dev --debug src/main.ts",
+    "lint": "eslint --fix '**/*.{ts,tsx}'",
+    "clean": "rm -rf ./dist",
+    "build": "npm run clean && npm run build:compile && npm run build:copy-assets",
+    "build:compile": "tsc -p tsconfig.json --outDir ./dist",
+    "build:copy-assets": "cp README.md ./dist && cp package.json ./dist",
+    "release": "standard-version --no-verify",
     "coverage": "nyc -r lcov -e .ts -x \"*[Tt]est.ts\" npm run test"
   },
   "repository": {
@@ -22,25 +25,28 @@
     "@commitlint/cli": "^15.0.0",
     "@commitlint/config-conventional": "^15.0.0",
     "@graphql-tools/schema": "^8.3.1",
-    "@kidsloop/eslint-config": "bitbucket:calmisland/kidsloop-eslint-config",
-    "@newrelic/apollo-server-plugin": "^1.1.1",
     "@types/cookie": "^0.4.1",
     "@types/ioredis": "^4.28.1",
     "@types/jsonwebtoken": "^8.5.5",
     "@types/newrelic": "^7.0.2",
     "@types/uuid": "^7.0.4",
     "@types/ws": "^7.2.5",
+    "@typescript-eslint/eslint-plugin": "^5.13.0",
+    "@typescript-eslint/parser": "^5.13.0",
     "cz-conventional-changelog": "^3.0.1",
+    "eslint": "^8.10.0",
+    "eslint-config-google": "^0.14.0",
+    "husky": "^7.0.4",
     "nyc": "^15.1.0",
+    "standard-version": "^9.3.2",
     "ts-loader": "^9.2.6",
     "ts-mockito": "^2.6.1",
     "ts-node": "^8.10.2",
     "ts-node-dev": "^1.1.8",
-    "typescript": "^4.3.5",
-    "webpack": "^5.58.2",
-    "webpack-cli": "^4.9.1"
+    "typescript": "^4.5.5"
   },
   "dependencies": {
+    "@newrelic/apollo-server-plugin": "^1.2.0",
     "apollo-server-express": "^3.5.0",
     "aws-sdk": "^2.926.0",
     "axios": "^0.21.4",
@@ -57,7 +63,6 @@
     "subscriptions-transport-ws": "^0.11.0",
     "typeorm": "^0.2.29",
     "uuid": "^8.1.0",
-    "webpack-node-externals": "^3.0.0",
     "ws": "^7.4.6"
   },
   "config": {
